verify largest extracted file
Allows for the scenario where a small unencrypted file is included in an archive of mostly encrypted files. E.g. a .txt file containing a password for the main content. In this scenario, the small file would give Extractor the impression that the whole archive wasn't password protected resulting in the largest encrypted file being extracted as a zero-length file.  This amendment verifies that the largest extract file is the correct size and therefore extracted successfully with the correct password.